Support SNAPSHOTs and fix dependency cache
We will no longer fail when a SNAPSHOT dependency is used. In order
to make sure that SNAPSHOTs are periodically re-resolved, a TTL of
1 day was added to the cache file so that the cache is invalidated
every 24 hours. The cache can be manually invalidated using
`smithy clean`.

When implementing a TTL, I noticed that caching wasn't actually
working and was always invalidated. That's fixed now by using the
correct timestamp comparisons, and several new test cases were added.

Closes #1850