Fix off-by-one issues in TokenTree and TreeCursor (#1891)
Fixes an issue in TokenTree where if the smithy file was missing a
trailing newline, any trees containing the last piece of text in
the file would have an end location of (0, 0). This happened because
the BR TreeType didn't append any tokens for the EOF case, so the
BR tree had no end location.

Fixes an issue in TreeCursor::findAt where if you tried to get the
tree at a location right at the start of a tree, you would get the
previous tree.

Tests were added for both of these cases in TreeCursorTest.