Clarify ordering of auth schemes in ServiceIndex (#1915)
Updates the docs for getEffectiveAuthSchemes to clarify that the
returned auth schemes will be in alphabetical order by shape id
if there is no `auth` trait present. This was always the case due
to the usage of TreeMap, but the documentation didn't explicitly
state it.

Tests were also updated to ensure this ordering.