Add validator for services with no auth trait (#1929)
Adds a validator that warns when there's a service shape that has
multiple authDefinition traits applied, but no auth trait. This
encourages the use of the auth trait so auth scheme priority is
being modeled explicitly. It also allows for tooling to look for
and use this validation event.

Documentation was also updated to explain that without the auth
trait, auth scheme ordering is alphabetical.

A test for the auth trait (auth-trait-must-target-service-schemes)
was updated to avoid getting the warning message from the new
validator.