Add transform that can remove invalid defaults
If a default trait is incompatible with the range trait of the member
or the member target, then the shape is essentially in a state where
the member is invalid: omit a value for the member at it is
automatically incompatible with the range trait of the member.

This transformer finds such cases and removes invalid default values.
It is likely that services with such a model did not intend for the
members to actually have a default value, and this typically only
happens with members that have a default value of `0`.